requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
plotly>=5.0.0
wordcloud>=1.8.0
bibtexparser>=1.4.0
openai>=0.27.0
tqdm>=4.60.0
python-dotenv>=0.19.0
Jinja2>=3.0.0

# For interactive visualizations
dash>=2.0.0
dash-bootstrap-components>=1.0.0

# For network visualizations (optional)
networkx>=2.6.0

# For dimensionality reduction (optional)
umap-learn>=0.5.0
scikit-learn>=0.24.0

# For testing
pytest>=6.0.0
pytest-cov>=2.0.0
pytest-mock>=3.0.0

# For code quality
black>=21.0
isort>=5.0.0
flake8>=3.9.0
mypy>=0.900
pre-commit>=2.0.0

# For documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0
